% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nombre_d_adjoints.R
\name{compter_nombre_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Compter le nombre d'adjoints}
\usage{
compter_nombre_d_adjoints(df)
}
\arguments{
\item{df}{Un DataFrame contenant les informations des élus municipaux. Il doit inclure une colonne Libellé.de.la.fonction.}
}
\value{
Un entier représentant le nombre total d'adjoints trouvés dans le DataFrame.
}
\description{
Cette fonction permet de compter le nombre d'adjoints présents dans un DataFrame contenant les élus conseillers municipaux.
}
\details{
La fonction recherche les lignes où Libellé.de.la.fonction contient le mot "adjoint".
}
