% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nombre_d_elu.R
\name{compter_nombre_d_elu}
\alias{compter_nombre_d_elu}
\alias{compter_nombre_d_elus}
\title{Compter le nombre d'élu}
\usage{
compter_nombre_d_elus(df)
}
\arguments{
\item{df}{Un data.frame contenant au minimum les colonnes Nom.de.l.élu, Prénom.de.l.élu et Date.de.naissance}
}
\value{
Un entier représentant le nombre d'élu
}
\description{
Cette fonction permet de compter le nombre d'élu pour le Dataframe: élus conseillés municipaux
}
