% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_commune.R
\name{creer_commune}
\alias{creer_commune}
\title{Création d'un objet spécifique pour les communes}
\usage{
creer_commune(df)
}
\arguments{
\item{df}{Un dataframe contenant les informations des élus municipaux, avec au minimum les colonnes suivantes :
\itemize{
\item \code{Code.de.la.commune} : Code de la commune.
\item \code{Libellé.de.la.commune} : Nom de la commune.
\item \code{Nom.de.l.élu} : Nom de l'élu.
\item \code{Prénom.de.l.élu} : Prénom de l'élu.
\item \code{Code.sexe} : Code du sexe de l'élu.
\item \code{Date.de.naissance} : Date de naissance de l'élu.
\item \code{Libellé.de.la.fonction} : Fonction de l'élu.
\item \code{Date.de.début.de.la.fonction} : Date de début de la fonction de l'élu.
\item \code{Code.nationalité} : Code de nationalité de l'élu.
}}
}
\value{
Un objet de type \code{commune} qui est une liste contenant des informations résumées pour une commune, telles que :
\itemize{
\item \code{Code.de.la.commune} : Code unique de la commune.
\item \code{Libellé.de.la.commune} : Nom de la commune.
\item \code{Nom.de.l.élu} : Liste des élus dans la commune.
\item \code{Prénom.de.l.élu} : Liste des prénoms des élus.
\item \code{Code.sexe} : Liste des codes sexe des élus.
\item \code{Date.de.naissance} : Liste des dates de naissance des élus.
\item \code{Libellé.de.la.fonction} : Liste des fonctions des élus.
\item \code{Date.de.début.de.la.fonction} : Liste des dates de début de fonction des élus.
\item \code{Code.nationalité} : Liste des codes de nationalité des élus.
}
}
\description{
Cette fonction permet de créer un objet de type commune à partir d'un dataframe contenant des informations sur les élus municipaux. Elle extrait les informations essentielles pour une commune donnée y compris les élus, leurs fonctions, et des informations de base les concernant.
}
