% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_de_donnees.R
\docType{data}
\name{df_Loire_Atlantique}
\alias{df_Loire_Atlantique}
\title{Jeu de données des élus du département de Loire Atlantique}
\format{
Un data.frame avec 4791 lignes et 16 colonnes :
\describe{
\item{Code.du.département}{Code du département selon la nomenclature officielle.}
\item{Libellé.du.département}{Nom du département}
\item{Code.de.la.collectivité.à.statut.particulier}{Code de la collectivité territoriale à statut particulier.}
\item{Libellé.de.la.collectivité.à.statut.particulier}{Nom de la collectivité territoriale à statut particulier.}
\item{Code.de.la.commune}{Code officiel de la commune}
\item{Libellé.de.la.commune}{Nom de la commune}
\item{Nom.de.l.élu}{Nom de l’élu}
\item{Prénom.de.l.élu}{Prénom de l’élu}
\item{Code.sexe}{Code indiquant le sexe de l’élu}
\item{Date.de.naissance}{Date de naissance de l’élu}
\item{Code.de.la.catégorie.socio.professionnelle}{Code représentant la catégorie socio-professionnelle de l’élu}
\item{Libellé.de.la.catégorie.socio.professionnelle}{Description de la catégorie socio-professionnelle}
\item{Date.de.début.du.mandat}{Date de début du mandat de l’élu}
\item{Libellé.de.la.fonction}{Fonction occupée par l’élu}
\item{Date.de.début.de.la.fonction}{Date de prise de fonction de l’élu}
\item{Code.nationalité}{Code indiquant la nationalité de l’élu}
}
}
\source{
Données récupérées depuis le site data.gouv \link{https://www.data.gouv.fr/fr/datasets/repertoire-national-des-elus-1/#/resources/d5f400de-ae3f-4966-8cb6-a85c70c6c24a}
}
\usage{
df_Loire_Atlantique
}
\description{
Ce dataset contient les informations sur les élus du département de Loire Atlantique, contenant par exemple leur nom, prénom, date de naissance, catégorie socio-professionnelle, etc.
}
\keyword{datasets}
