% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Générer un rapport Quarto pour une commune et un département}
\usage{
generer_rapport(commune, departement, output)
}
\arguments{
\item{commune}{Un code de commune sous forme de chaîne de caractères (ex : "44109").}

\item{departement}{Un code de département sous forme de chaîne de caractères (ex : "44").}

\item{output}{Chemin de sortie du fichier HTML généré, sous forme de chaîne de caractères.}
}
\value{
un fichier html
}
\description{
Cette fonction génère un rapport HTML à partir d'un fichier Quarto. Elle utilise les codes de la commune et du département spécifiés, puis enregistre le rapport dans un dossier html
}
\examples{
\dontrun{
generer_rapport(
 commune = "44330",
 departement = "44",
 output = "rapport_la_chapelle.html"
)
}

}
