% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_departement.R
\name{plot.departement}
\alias{plot.departement}
\alias{plot_departement}
\title{Visualisation de la répartition des élus par catégorie socio-professionnelle pour un département}
\usage{
plot_departement(df)
}
\arguments{
\item{df}{Un dataframe contenant les informations des élus municipaux d'un seul département.
Il doit inclure au minimum les colonnes suivantes :
\itemize{
\item \code{Libellé.du.département} : Nom du département.
\item \code{Libellé.de.la.commune} : Nom de la commune.
\item \code{Code.de.la.catégorie.socio.professionnelle} : Code représentant la catégorie socio-professionnelle des élus.
}}
}
\value{
Un graphique en barres affichant le nombre d'élus par code professionnel pour le département sélectionné.
}
\description{
Cette fonction génère un graphique en barres représentant la répartition des élus municipaux en fonction de leur catégorie socio-professionnelle pour un département donné.
}
